Restore current DB migrations and renamed the license export one
