+++ migration/src/lib.rs
@@ -1,15 +1,14 @@
 pub use sea_orm_migration::prelude::*;
 
-mod m0000010_init;
-mod m0000020_add_sbom_group;
-mod m0000080_create_license_export;
-mod m0000970_alter_importer_add_heartbeat;
-
 #[cfg(feature = "ai")]
 pub mod ai;
 #[cfg(feature = "ai")]
 mod ai_m0000010_create_conversation;
+mod m0000010_init;
+mod m0000020_add_sbom_group;
 mod m0000030_perf_adv_vuln;
+mod m0000040_create_license_export;
+mod m0000970_alter_importer_add_heartbeat;
 
 pub struct Migrator;
 
@@ -18,10 +17,10 @@ impl MigratorTrait for Migrator {
     fn migrations() -> Vec<Box<dyn MigrationTrait>> {
         vec![
             Box::new(m0000010_init::Migration),
+            Box::new(m0000970_alter_importer_add_heartbeat::Migration),
             Box::new(m0000020_add_sbom_group::Migration),
             Box::new(m0000030_perf_adv_vuln::Migration),
-            Box::new(m0000080_create_license_export::Migration),
-            Box::new(m0000970_alter_importer_add_heartbeat::Migration),
+            Box::new(m0000040_create_license_export::Migration),
         ]
     }
 }
